index on master: 22bbac6 + Added is_mouse_inside() to bee/game/input.cpp for testing whether the mouse is within an instance + Added a new version of draw_fast() to the Font class to enable drawing with an rgba_t color + Made sure that the game variable for new objects is set in Room::add_instance() + Added the variable has_draw_failed to the Sprite class so that failed draws from unloaded images aren't spammed to the console * Fully fixed Makefile

+++ bee/resource_structures/ext/instancedata.cpp
@@ -27,8 +27,8 @@ int BEE::InstanceData::init(int new_id, Object* new_object, int new_x, int new_y
 	subimage_time = SDL_GetTicks();
 	depth = object->get_depth();
 
-	x = (float)new_x;
-	y = (float)new_y;
+	x = (double)new_x;
+	y = (double)new_y;
 	xprevious = x;
 	yprevious = y;
 	xstart = x;
